+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add `Cache#clear()` method
 * Add `Function.postInherit(task)`, which will execute the task after the class has been inherited
 * Add `Blast.parseClassPath(path)` method
+* Add `WeakValueSet` class
 
 ## 0.8.18 (2024-01-19)